backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
bcrypt==3.2.2

